Added cancel token validator
